backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
aiosqlite
python-multipart
httpx
pyyaml
google-generativeai
passlib[bcrypt]
python-jose[cryptography]
